* Move the GDS (in particular the dstore) locks up to the `_mdxcbfunc`
   level to avoid the over head of locking and unlocking for each key
   set in the modex data.