@ibm-ompi 2018-10-10
The IBM CI (GNU Compiler) build failed! Please review the log, linked below.

 Gist: https://gist.github.com/0413a092764148806bcd8338cf78eac4