@ibm-ompi 2018-10-10
The IBM CI (XL Compiler) build failed! Please review the log, linked below.

 Gist: https://gist.github.com/eff2b126c21440eb61b7cca4168633c8