@ibm-ompi 2018-10-10
The IBM CI (GNU Compiler) build failed! Please review the log, linked below.

 Gist: https://gist.github.com/7730a12150106828b55e7c06dd819b8f