@jjhursey 2018-10-10
We are performance tuning at full scale with high PPN, and I noticed that the accumulation of the locking overhead inside the new dstore on the server side was considerable when processing a full modex. At smaller scales, it's hardly noticeable (10's to 100's of usec of accumulation), but at large scale it was 10's of seconds.

In our scenario we are doing a full modex and each process is contributing 1-2 key/value pairs. Each node has 168 PPN. The numbers below are using `hash,ds21`.
The main loop in `pmix_server.c:_mdxcbfunc` will have an entry for each host. The time quoted below is the time to process 1 of those per-host entries.
```
Without this PR:
0.012326 usec x nodes
  -  6.310 sec @  512
  - 56.798 sec @ 4608

With this PR:
0.004025 usec x nodes
  -  2.06 sec @  512
  - 18.547 sec @ 4608
```

The numbers above were projected from small scale measurement to the node counts noted. The performance benefit at 512 nodes was later confirmed on a scale machine.