@karasevb 2018-10-10
Our Jenkins has issues (`not ok`):
```
18:07:33 Checking with ds21,hash:
18:07:33 1..14
18:07:33 not ok 1 blocking fence w/ data all (./pmix_test -n 4 --ns-dist 3:1 --fence "[db | 0:0-2;1:3]" -o $OUTDIR/out)
18:07:33 not ok 2 blocking fence w/ data all (./pmix_test -n 4 --ns-dist 3:1 --fence "[db | 0:;1:3]" -o $OUTDIR/out)
18:07:33 not ok 3 blocking fence w/ data all (./pmix_test -n 4 --ns-dist 3:1 --fence "[db | 0:;1:]" -o $OUTDIR/out)
18:07:33 ok 4 non-blocking fence w/o data (./pmix_test -n 4 --ns-dist 3:1 --fence "[0:]" -o $OUTDIR/out)
18:07:33 ok 5 blocking fence w/ data (./pmix_test -n 4 --ns-dist 3:1 --fence "[b | 0:]" -o $OUTDIR/out)
18:07:33 ok 6 non-blocking fence w/ data (./pmix_test -n 4 --ns-dist 3:1 --fence "[d | 0:]" --noise "[0:0,1]" -o $OUTDIR/out)
18:07:33 ok 7 blocking fence w/ data on the same nspace (./pmix_test -n 4 --job-fence -c -o $OUTDIR/out)
18:07:33 ok 8 blocking fence w/o data on the same nspace (./pmix_test -n 4 --job-fence -o $OUTDIR/out)
18:07:33 ok 9 publish (./pmix_test -n 2 --test-publish -o $OUTDIR/out)
18:07:33 ok 10 spawn (./pmix_test -n 2 --test-spawn -o $OUTDIR/out)
18:07:33 ok 11 connect (./pmix_test -n 2 --test-connect -o $OUTDIR/out)
18:07:33 ok 12 resolve peers (./pmix_test -n 5 --test-resolve-peers --ns-dist "1:2:2" -o $OUTDIR/out)
18:07:33 ok 13 key replacement (./pmix_test -n 5 --test-replace 100:0,1,10,50,99 -o $OUTDIR/out)
18:07:33 ok 14 local store (./pmix_test -n 5 --test-internal 10 -o $OUTDIR/out)
```
Let me check what is wrong.