@sam6258 2018-10-11
@karasevb Is the ds21 component designed to be re-entrant?

It looks like that is the key difference with this PR's CI failures with ds21.

With ds12 we see the server perform `pmix_ds12_lock_wr_get` without a release of the lock:

`[c685f10n17:130869] Locking namespace 0x200004000a90, id is smoky_nspace-0
[c685f10n17:130869] ns_map->tbl_idx = 0
[c685f10n17:130869] obtain lock in pmix_common_dstor_acquire_ns_lock
[c685f10n17:130869] pmix_ds12_lock_wr_get
[c685f10n17:130869] PMIX_GDS_ACQUIRE_NS_LOCK rc is 0
[c685f10n17:130869] Locking namespace 0x200004010b40, id is smoky_nspace-1
[c685f10n17:130869] ns_map->tbl_idx = 0
[c685f10n17:130869] obtain lock in pmix_common_dstor_acquire_ns_lock
[c685f10n17:130869] pmix_ds12_lock_wr_get
[c685f10n17:130869] PMIX_GDS_ACQUIRE_NS_LOCK rc is 0`

With ds21, we hang on first re-entry:

`[c685f10n17:131854] Locking namespace 0x200004000a90, id is smoky_nspace-0
[c685f10n17:131854] ns_map->tbl_idx = 0
[c685f10n17:131854] obtain lock in pmix_common_dstor_acquire_ns_lock
[c685f10n17:131854] pmix_ds21_lock_wr_get
[c685f10n17:131854] PMIX_GDS_ACQUIRE_NS_LOCK rc is 0
[c685f10n17:131854] Locking namespace 0x200004010c50, id is smoky_nspace-1
[c685f10n17:131854] ns_map->tbl_idx = 0
[c685f10n17:131854] obtain lock in pmix_common_dstor_acquire_ns_lock`