@rhc54 2018-10-11
IIRC, none of the GDS components are intended or designed for re-entrant use. We utilize the event library to serialize data access. The locks here are solely intended to coordinate between the server and its clients, not between threads in either entity.