@jjhursey 2018-10-11
Let's refactor the 'acquire/release_ns' locks to take a list of namespaces. Move the loop from pmix_server.c to inside the dstore components. Then they are resolve the subset of `ns_map`s that need locking to prevent it from trying to grab the same lock twice.

I think that would give us the control we need at that layer, right?