@sam6258 2018-10-11
There is one problem with that solution. Once you enter the `PMIX_GDS_ACQUIRE_NS_LOCK` macro, you need to look at each namespace's personality (compat) to get the ds component that it is using. You would need to aggregate multiple lists (one list for each component), and then call corresponding `acquire_ns_lock` function. After all that is said and done, that seems like it would be pretty hacky.

Another solution could be to keep the current structure of function calls, but in the dstore base, add logic that tracks outstanding locks, so we don't re-enter a lock. Perhaps an `active_lock` value could be added to `ns_map_data_t`