@karasevb 2018-10-12
@sam6258  keeping the `active_lock` in `ns_map_data_`t is not a good idea. Since `ns_map_data_t` is an entry for tracking of the nspace session data, if someone sets the `active_lock` it will not propagate to others nspace `ns_map_data_t` entries with the same session. There will be the same deadlock.
The first way would be better.