@jjhursey 2018-10-12
@karasevb Scott refined the locking in the second commit on this PR which seems to address your concern (and cleanup the 'holding_lock' field to the gds modex operation).

I think this is good to go for re-review.