@artpol84 2018-10-13
@jjhursey this looks good.
As the new writer lock is heavy we need to reduce the amount of locks. I was thinking about this optimizatio, but forgot to schedule its implementation. Thanks for capturing this.