@artpol84 2018-10-15
This was sounding very familiar and I finally realized why.
We moved the lock up for the job-level info store: a38d149311e9d17a068f46215d2849ab08debbbe. But we missed this modex store.
@jjhursey how do those changes go together? Are they independent or you are doing this for the job-level as well?