@artpol84 2018-10-15
So looking on the way we deal with job-level info I was thinking that we can aproach this problem from the other angle. Exactly the same way we do with job-level info: just pass the whole thing to the GDB and let it store the full set of keys allowing coarse-grained locking as opposed to storing each key individually.
This could be lighter solution requiring one additional callback to GDS and will allow to avoid exposure of locking semantic to the upper layers.