@jjhursey 2018-10-17
I just pushed the squash, and a documentation update to `gds.h` instead of renaming the operation. The "lock" already expresses exclusivity, so just make it clear in the header that it's an exclusive lock on the segment(s) associated with that namespace.

So I think this is good to go once CI is done.